+++ IQDMPDF/_version.py
@@ -1,3 +1,3 @@
 __author__ = "Dan Cutright"
-__version__ = "0.2"
-__release__ = "0.2"
+__version__ = "0.2.1"
+__release__ = "0.2.1"
